+++ Forge/src/main/java/com/christofmeg/brutalharvest/client/data/BrutalLanguageProvider.java
@@ -36,6 +36,10 @@ public class BrutalLanguageProvider extends LanguageProvider {
             add(CommonConstants.MOD_ID + "." + "advancement" + "." + "rotten_tomatoes" + ".desc", "Throw a tomato at a villager");
             add(CommonConstants.MOD_ID + "." + "advancement" + "." + "grim_reaper", "Grim Reaper");
             add(CommonConstants.MOD_ID + "." + "advancement" + "." + "grim_reaper" + ".desc", "Craft a scythe");
+            add(CommonConstants.MOD_ID + "." + "advancement" + "." + "corn_seeds", "Colonel Cornelius Cornwall");
+            add(CommonConstants.MOD_ID + "." + "advancement" + "." + "corn_seeds" + ".desc", "Plant a Corn Seed (Kernel)");
+            add(CommonConstants.MOD_ID + "." + "advancement" + "." + "corn", "IT'S CORN");
+            add(CommonConstants.MOD_ID + "." + "advancement" + "." + "corn" + ".desc", "Obtain some corn");
 
             add("sounds." + CommonConstants.MOD_ID + "." + "tomato_splat", "Tomato Splat");