+++ Forge/changelog.md
@@ -4,3 +4,4 @@
 * Added Unripe Strawberry food item
 * Added Strawberry Seeds
 * Added Sugar Beets food properties
+* Added Advancement for Corn and Corn Seeds