+++ Forge/changelog.md
@@ -1,2 +1,5 @@
 1.7.2:
 * Update Strawberry Seed texture
+* Added Cucumber Crop textures
+* Added Coffee Crop textures
+* Added textures for future blocks and items